+++ app/nft/burn.ts
@@ -3,7 +3,7 @@ import { abi } from "@zetachain/example-contracts/abi/omnichain/NFT.sol/NFT.json
 import { ethers } from "ethers"
 import { useAccount } from "wagmi"
 
-import { useEthersSigner } from "@/lib/ethers"
+import { useEthersSigner } from "@/hooks/useEthersSigner"
 import { AppContext } from "@/app/index"
 
 import { useFetchNFTs } from "./fetchNFTs"
