+++ app/nft/transfer.ts
@@ -2,7 +2,7 @@ import { abi } from "@zetachain/example-contracts/abi/omnichain/NFT.sol/NFT.json
 import { ethers } from "ethers"
 import { useAccount } from "wagmi"
 
-import { useEthersSigner } from "@/lib/ethers"
+import { useEthersSigner } from "@/hooks/useEthersSigner"
 
 import { useFetchNFTs } from "./fetchNFTs"
 import { useNFT } from "./useNFT"
